BUG: Fix help if no input provided. - Fix help if no input provided. - Remove redundant code. - Use common click options for consistency.
